[HUDI-9144][part4] Add tests for RowData upsert writer for MOR table
