+++ hudi-flink-datasource/hudi-flink/src/test/java/org/apache/hudi/sink/TestWriteMergeOnRead.java
@@ -41,6 +41,8 @@ public class TestWriteMergeOnRead extends TestWriteCopyOnWrite {
   @Override
   protected void setUp(Configuration conf) {
     conf.setBoolean(FlinkOptions.COMPACTION_ASYNC_ENABLED, false);
+    // test legacy stream write function
+    conf.setBoolean(FlinkOptions.INSERT_ROWDATA_MODE_ENABLED, false);
   }
 
   @Test
