+++ hudi-flink-datasource/hudi-flink/src/test/java/org/apache/hudi/sink/utils/TestWriteBase.java
@@ -639,6 +639,6 @@ public class TestWriteBase {
   }
 
   protected HoodieTableType getTableType() {
-    return HoodieTableType.COPY_ON_WRITE;
+    return HoodieTableType.MERGE_ON_READ;
   }
 }
