+++ hudi-flink-datasource/hudi-flink/src/test/java/org/apache/hudi/utils/TestData.java
@@ -33,10 +33,12 @@ import org.apache.hudi.common.util.Option;
 import org.apache.hudi.config.HoodieWriteConfig;
 import org.apache.hudi.configuration.FlinkOptions;
 import org.apache.hudi.configuration.OptionsResolver;
+import org.apache.hudi.sink.utils.BucketRowDataStreamWriteFunctionWrapper;
 import org.apache.hudi.sink.utils.BucketStreamWriteFunctionWrapper;
 import org.apache.hudi.sink.utils.BulkInsertFunctionWrapper;
 import org.apache.hudi.sink.utils.ConsistentBucketStreamWriteFunctionWrapper;
 import org.apache.hudi.sink.utils.InsertFunctionWrapper;
+import org.apache.hudi.sink.utils.RowDataStreamWriteFunctionWrapper;
 import org.apache.hudi.sink.utils.StreamWriteFunctionWrapper;
 import org.apache.hudi.sink.utils.TestFunctionWrapper;
 import org.apache.hudi.storage.HoodieStorage;
@@ -577,10 +579,12 @@ public class TestData {
       if (OptionsResolver.isConsistentHashingBucketIndexType(conf)) {
         return new ConsistentBucketStreamWriteFunctionWrapper<>(basePath, conf);
       } else {
-        return new BucketStreamWriteFunctionWrapper<>(basePath, conf);
+        return OptionsResolver.supportRowDataAppend(conf)
+            ? new BucketRowDataStreamWriteFunctionWrapper<>(basePath, conf) : new BucketStreamWriteFunctionWrapper<>(basePath, conf);
       }
     } else {
-      return new StreamWriteFunctionWrapper<>(basePath, conf);
+      return OptionsResolver.supportRowDataAppend(conf)
+          ? new RowDataStreamWriteFunctionWrapper<>(basePath, conf) : new StreamWriteFunctionWrapper<>(basePath, conf);
     }
   }
